+++ android/src/main/java/com/applovin/reactnative/AppLovinMAXNativeAdView.java
@@ -4,6 +4,7 @@ import android.content.Context;
 import android.text.TextUtils;
 import android.view.View;
 import android.view.ViewGroup;
+import android.widget.Button;
 import android.widget.ImageView;
 
 import com.applovin.mediation.MaxAd;
@@ -34,6 +35,13 @@ public class AppLovinMAXNativeAdView
         extends ReactViewGroup
         implements MaxAdRevenueListener, View.OnLayoutChangeListener, ViewGroup.OnHierarchyChangeListener
 {
+    private static final int TITLE_LABEL_TAG          = 1;
+    private static final int MEDIA_VIEW_CONTAINER_TAG = 2;
+    private static final int ICON_VIEW_TAG            = 3;
+    private static final int BODY_VIEW_TAG            = 4;
+    private static final int CALL_TO_ACTION_VIEW_TAG  = 5;
+    private static final int ADVERTISER_VIEW_TAG      = 8;
+
     private final ReactContext      reactContext;
     @Nullable
     private       MaxNativeAdLoader adLoader;
@@ -173,7 +181,7 @@ public class AppLovinMAXNativeAdView
         }
 
         view.setClickable( true );
-
+        view.setTag( TITLE_LABEL_TAG );
         clickableViews.add( view );
     }
 
@@ -189,7 +197,7 @@ public class AppLovinMAXNativeAdView
         }
 
         view.setClickable( true );
-
+        view.setTag( ADVERTISER_VIEW_TAG );
         clickableViews.add( view );
     }
 
@@ -205,7 +213,7 @@ public class AppLovinMAXNativeAdView
         }
 
         view.setClickable( true );
-
+        view.setTag( BODY_VIEW_TAG );
         clickableViews.add( view );
     }
 
@@ -222,7 +230,22 @@ public class AppLovinMAXNativeAdView
 
         view.setClickable( true );
 
-        clickableViews.add( view );
+        // Some adapters, like Google, expect a Button widget for CTA to be clickable
+        if ( view instanceof ViewGroup )
+        {
+            Button button = new Button( reactContext );
+            button.setAlpha( 0 );
+            ( (ViewGroup) view ).addView( button );
+            sizeToFit( button, view );
+
+            button.setTag( CALL_TO_ACTION_VIEW_TAG );
+            clickableViews.add( button );
+        }
+        else
+        {
+            view.setTag( CALL_TO_ACTION_VIEW_TAG );
+            clickableViews.add( view );
+        }
     }
 
     public void setIconView(final int tag)
@@ -235,6 +258,7 @@ public class AppLovinMAXNativeAdView
         }
 
         view.setClickable( true );
+        view.setTag( ICON_VIEW_TAG );
         clickableViews.add( view );
 
         MaxNativeAdImage icon = nativeAd.getNativeAd().getIcon();
@@ -278,6 +302,7 @@ public class AppLovinMAXNativeAdView
             return;
         }
 
+        view.setTag( MEDIA_VIEW_CONTAINER_TAG );
         clickableViews.add( view );
 
         view.addOnLayoutChangeListener( this );
@@ -378,7 +403,7 @@ public class AppLovinMAXNativeAdView
                 // LINE, where the actual ad contents are loaded after `mediaView` is sized.
                 if ( mediaView != null )
                 {
-                    sizeToFit( mediaView, (ReactViewGroup) mediaView.getParent() );
+                    sizeToFit( mediaView, (View) mediaView.getParent() );
                 }
 
                 isLoading.set( false );